buildsystem: Build as freestanding, not hosted library
libopencm3 is designed to run as a bare-metal library, and as such, we
need to tell the compiler this by passing the "-ffreestanding" flag.

Thus has ugly side-effects. For one, we lose stdlib, from which we use
strlen() and memcpy(). This poses an interesting egg-and chicken
problem:

Do we stand above or below the C library?

Common sense dictates that we stand below the C library, in which case
we'll need to somehow provide our own versions of these functions.
Most other firmware projects (coreboot, uboot, etc) already do, so
this is not so terrible. Unfortunately, we run in the same namespace
as a potential C library, so this is bount to cause name collisions.
Renaming our memcpy to anything else is plain ID10Tic, so we could
make our implementations weak. That sadly makes us vulnerable to bugs
in the memcpy of whoever else provides it. IT'S A CLUSTERFUCK!!!!!

The fact that we have compiled until now is probably the result of
having always used a specially tailored toolchain, which did some of
the work for us and handled some of our assumptions.

Nowadays, distros are packaging different flavors of arm-*-gcc, which
do not treat us so lavishly. The behavior of us fucking up while other
bare metal firmwares do not, seems to be fairly consistent across
distros and flavors of gcc (vanilla, codesourcery, etc ...). We're
fucking up somewhere, and we need to figure out how to pull our dicks
out.

P.S. DO NOT FUCKING MERGE